Implement routing and navigation structure with React Router
- Added routing capabilities using 'react-router-dom'.
- Created Home, Navigation, and Shop components for structured navigation.
- Updated App component to include routes for Home and Shop.
- Introduced styles for navigation layout in navigation.styles.scss.
- Updated index.scss for global styles affecting links and box-sizing.
