+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "6",
     "sass": "^1.86.0"
   },
   "devDependencies": {
