+++ src/App.jsx
@@ -1,35 +1,21 @@
-import Directory from './components/directory/directory.component'
+import { Route, Routes } from 'react-router-dom'
+
+import Home from './routes/home/home.component'
+import Navigation from './routes/home/navigation/navigation.component'
+
+const Shop = () => {
+  return <h2>Im the shop page</h2>
+}
 
 export const App = () => {
-  const categories = [
-    {
-      id: 1,
-      title: 'hats',
-      imageUrl: 'https://i.ibb.co/cvpntL1/hats.png',
-    },
-    {
-      id: 2,
-      title: 'jackets',
-      imageUrl: 'https://i.ibb.co/px2tCc3/jackets.png',
-    },
-    {
-      id: 3,
-      title: 'sneakers',
-      imageUrl: 'https://i.ibb.co/0jqHpnp/sneakers.png',
-    },
-    {
-      id: 4,
-      title: 'womens',
-      imageUrl: 'https://i.ibb.co/GCCdy8t/womens.png',
-    },
-    {
-      id: 5,
-      title: 'mens',
-      imageUrl: 'https://i.ibb.co/R70vBrQ/men.png',
-    },
-  ]
-
-  return <Directory categories={categories} />
+  return (
+    <Routes>
+      <Route path="/" element={<Navigation />}>
+        <Route index element={<Home />} />
+        <Route path="/shop" element={<Shop />} />
+      </Route>
+    </Routes>
+  )
 }
 
 export default App
