+++ src/routes/home/navigation/navigation.component.jsx
@@ -1,0 +1,23 @@
+import { Outlet, Link } from 'react-router-dom'
+
+import './navigation.styles.scss'
+
+const Navigation = () => {
+  return (
+    <>
+      <div className="navigation">
+        <Link className="logo-container" to="/">
+          Logo
+        </Link>
+        <div className="nav-links-container">
+          <Link className="nav-link" to="/shop">
+            SHOP
+          </Link>
+        </div>
+      </div>
+      <Outlet />
+    </>
+  )
+}
+
+export default Navigation
